LayoutTests/imported/w3c/web-platform-tests/css/css-text/line-breaking/line-breaking-028-expected.txt

X፡!
X !
X !
X !

PASS U+0x1361 (BA) and U+0x0021 (EX)
PASS U+0x1680 (BA) and U+0x0021 (EX)
PASS U+0x2009 (BA) and U+0x0021 (EX)
PASS U+0x205F (BA) and U+0x0021 (EX)
